backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
google-cloud-firestore==2.16.0
google-cloud-secret-manager==2.20.2
google-cloud-aiplatform==1.66.0
requests==2.32.3
pydantic==2.8.2
